requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cryptography==39.0.2
cymem==2.0.7
doc2text==0.2.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
future==0.18.3
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
mime==0.1.0
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
pdfminer.six==20221105
Pillow==9.4.0
preshed==3.0.8
pycparser==2.21
pydantic==1.10.6
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
